Fixed Poison.SyntaxError when receiving 204 status
I was using hound's navigate_to helper on BrowserStack's Amazon
KindleFire 2 device, and was receiving a 204 No Content response after a
POST /session/<session_id>/url request that changes the page. Since the
request succeeded, it seems Hound should be able to handle the response
without throwing the following error.

Unexpected end of input at position 0
  (poison) lib/poison/parser.ex:55: Poison.Parser.parse!/2
  (poison) lib/poison.ex:83: Poison.decode!/2
  (hound) lib/hound/response_parser.ex:29: Hound.ResponseParser.parse/5
  (hound) lib/hound/request_utils.ex:48: Hound.RequestUtils.handle_response/3

This change just skips the response parsing if the status code is 204
and the raw content is empty (which it should be).